Prevent exception with browser autocompletion
I got an exception with "Google smart lock" in chrome. In single page application, if there is hidden login page where chrome automatically fills user's data. Then keyboard is shown because of autocompletion and we get an exception in $keyboard.caret() function "Cannot read length of null".

With this changes keyboard is still opened but at least there is no exception.